Move extensions to their own namespaced files
Implements part of https://github.com/ruby/csv/issues/5

This does not change any behavior of the csv gem.
