+++ lib/ext/string.rb
@@ -1,0 +1,9 @@
+class String # :nodoc:
+  # Equivalent to CSV::parse_line(self, options)
+  #
+  #   "CSV,data".parse_csv
+  #     #=> ["CSV", "data"]
+  def parse_csv(**options)
+    CSV.parse_line(self, options)
+  end
+end
